Transition to insulating sqlalchemy objects entirely
In order to allow the most portability for web interfaces, no sqlalchemy instances will be exposed to frontends.

Instead, dictionaries will be passed to and accepted from frontends, and this backend will do whatever it needs to interface with sqlalchemy.

Note that this backend will need to be expanded to allow for all that nice stuff that sqlalchemy lets us do, like append related records to relationship lists.

To facilitate this, relationships are now eagerly loaded.